docs(samples): create model sample
